✅ : ensure authorized http client decorator deletes cache on forbidden

+++ lib/main/decorators/authorized_http_client_decorator.dart
@@ -29,11 +29,13 @@ class AuthorizedHttpClientDecorator implements HttpClient {
         headers: authorizedHeaders,
       );
       return data;
-    } on HttpError {
-      rethrow;
-    } on Exception {
-      await deleteSecureCacheStorage.deleteSecure('token');
-      throw HttpError.forbidden;
+    } catch (error) {
+      if (error is HttpError && error != HttpError.forbidden) {
+        rethrow;
+      } else {
+        await deleteSecureCacheStorage.deleteSecure('token');
+        throw HttpError.forbidden;
+      }
     }
   }
 }
